refactor: split off functions to get direct origin dependencies from provider (#7693)
